requirements.prod.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
anyio==4.6.2.post1
attrs==24.2.0
Authlib==1.3.2
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
decorator==5.1.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
filelock==3.12.4
frozenlist==1.5.0
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hiredis==3.0.0
hpack==3.0.0
hstspreload==2024.12.1
httpcore==0.9.1
httptools==0.6.4
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.1.3
pandas==2.2.3
platformdirs==4.3.6
psutil==6.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.2.0
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
websockets==14.0
xxhash==3.5.0
yarl==1.17.1
